refactor(utils): rename utils.ts to functions.ts
